docker2smtp/requirements.txt
docker
paho-mqtt
